E17: symlink fixes
when copying symlinks, it creates a symlink to the
destination but with the name of the pointed file/folder.

It causes problems, ex: we can't copy two symlinks
pointing to the same file, because they will have the
same name.

Also when copying, if it's a link, we fill the destination
in the struct task, then, after we check overwrite with
_e_fm_op_handle_overwrite()

Here is a patch which corrects this.

Patch by Maxime Villard (rustyBSD), modified a bit by me (add 2 free() + formatting)


git-svn-id: http://svn.enlightenment.org/svn/e/trunk/e@74264 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33