From: Maxime Villard <rustyBSD@gmx.fr>
There is a problem with the realpath() call. When moving
a symlink, realpath() gets the path of the pointed file,
and the name assigned to the copied link is the name of
this file.

So we shouldn't use realpath(), but I don't know an equivalent
which doesn't take care of symlinks.

Here is an example patch.


git-svn-id: http://svn.enlightenment.org/svn/e/trunk/e@76721 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33