e: don't drop cache if icon is animated.
We want to use the scale cache and drop the source icon
for all static icon on screen. But for the few of them that
do resize and animate, we need to keep that data or it will
slow a us down a lot. So now we play nicely with our SCALE_HINT.


git-svn-id: http://svn.enlightenment.org/svn/e/trunk/e@78419 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33