E entry: force disable thumbscroll.
We need that disabled, as currently mouse grabbing breaks if that's not
set. That's what you get when you have scrolling and selection in the same
place. We can just use selection for scrolling or fix the issue.
The fix would probably require using the ON_SCROLL flag instead of
the ON_HOLD and actually handle it correctly.

git-svn-id: http://svn.enlightenment.org/svn/e/trunk/e@79359 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33