Add function for a monitor to apply it's changes.
NB: This is needed because apparently the e_randr serialization code
works DIRECTLY on the E_Randr_Info structures (E_Randr_Output_Info,
etc, etc). Sooo, in order to save and restore a randr config, we need
for the monitor to update those structures First before we tell
e_randr to save things else it will save with incorrect values.

NB: This is the START of save/restore configuration work.

Signed-off-by: Christopher Michael <cp.michael@samsung.com>

git-svn-id: http://svn.enlightenment.org/svn/e/trunk/e@81161 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33