Revert "e/connman: just update popup list at each 5 secs"
This reverts commit r79842 for the exact reason stated in the commit
log. First I thought it would not affect the usability, but it indeed
did. And it's a bit worse than having to wait 5 seconds to see the state
changes:  you never see them, except if you close and open the popup
again. This could be "fixed" by setting up the timer in
_econnman_popup_update() inside "if (inst->popup_locked)", but we would
still not see the transitions, only the final state.

A proper solution would be to first separate the service state changes
from the list updates, then do something like what was proposed in
r79842. Revert it while we don't have a solution that doesn't break the
more common cases of clicking in a service and see it connecting.



git-svn-id: http://svn.enlightenment.org/svn/e/trunk/e@81407 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33